+++ NEWS.md
@@ -1,11 +1,15 @@
 # raustats 0.1.0.9000
+---------------------------------------------------------------------
 
 ## Bug fixes:
-* First bug fix
+* Fixed bug in `abs_cat_stats` function to avoid downloading the files multiple
+  times and apply `abs_cat_unzip` function only to compressed files.
 
 ## Changes:
 * Added `na.rm` argument to `abs_cat_stats` and `abs_read_tss` functions.
 
 
 # raustats 0.1.0
+---------------------------------------------------------------------
+
 * Initial package release